@@ -1,10 +1,6 @@
-# encoding: utf-8
-
 """
 .. codeauthor:: Tsuyoshi Hombashi <tsuyoshi.hombashi@gmail.com>
 """
-
-from __future__ import absolute_import, unicode_literals
 
 
 PROGRAM_NAME = "sqlitebiter"
@@ -14,7 +10,7 @@ IPYNB_FORMAT_NAME_LIST = ["ipynb"]
 TABLE_NOT_FOUND_MSG_FORMAT = "convertible table not found in {}"
 
 
-class ExitCode(object):
+class ExitCode:
     SUCCESS = 0
     FAILED_LOADER_NOT_FOUND = 1
     FAILED_CONVERT = 2
