@@ -1,20 +1,16 @@
-# encoding: utf-8
-
 """
 .. codeauthor:: Tsuyoshi Hombashi <tsuyoshi.hombashi@gmail.com>
 """
 
-from __future__ import absolute_import, unicode_literals
 
 import re
 
 import msgfy
 import pytablereader as ptr
-import six
 from simplesqlite import SQLiteTableDataSanitizer
 
 
-class DictConverter(object):
+class DictConverter:
     @property
     def converted_table_name_set(self):
         return self.__converted_table_name_set
@@ -24,7 +20,7 @@ class DictConverter(object):
         self.__table_creator = table_creator
         self.__index_list = index_list
         self.__source_info = source_info
-        self.__converted_table_name_set = set([])
+        self.__converted_table_name_set = set()
 
     def to_sqlite_table(self, data, key_list):
         if not data:
@@ -33,7 +29,7 @@ class DictConverter(object):
         root_maps = {}
 
         for key, v in data.items():
-            if isinstance(v, (six.text_type, float) + six.integer_types) or v is None:
+            if isinstance(v, (str, float) + (int,)) or v is None:
                 root_maps[key] = v
                 continue
 
@@ -70,7 +66,7 @@ class DictConverter(object):
         return "_".join(key_list)
 
     def __convert(self, table_data):
-        self.__logger.debug("loaded tabledata: {}".format(six.text_type(table_data)))
+        self.__logger.debug("loaded tabledata: {}".format(str(table_data)))
 
         sqlite_tabledata = SQLiteTableDataSanitizer(table_data).normalize()
         self.__table_creator.create(
