@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 import sphinx_rtd_theme
 
 from sqlitebiter import __author__, __copyright__, __name__, __version__
@@ -214,8 +212,8 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'sqlitebiter.tex', u'sqlitebiter Documentation',
-     u'Tsuyoshi Hombashi', 'manual'),
+    (master_doc, 'sqlitebiter.tex', 'sqlitebiter Documentation',
+     'Tsuyoshi Hombashi', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -244,7 +242,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'sqlitebiter', u'sqlitebiter Documentation',
+    (master_doc, 'sqlitebiter', 'sqlitebiter Documentation',
      [author], 1)
 ]
 
@@ -258,7 +256,7 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'sqlitebiter', u'sqlitebiter Documentation',
+    (master_doc, 'sqlitebiter', 'sqlitebiter Documentation',
      author, 'sqlitebiter', 'One line description of project.',
      'Miscellaneous'),
 ]
